# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": "API teste"
  },
  "host": "localhost:3000",
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Ambiente de desenvolvimento"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/user": {
      "post": {
        "tags": [
          "user"
        ],
        "description": "Endpoint para criar um novo usuário",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "Informações necessárias para criar um novo usuário",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "user@example.com"
                },
                "nome": {
                  "type": "string",
                  "example": "João da Silva"
                },
                "senha": {
                  "type": "string",
                  "example": "senha123"
                },
                "permission": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso"
          },
          "422": {
            "description": "Erro de validação. Campos obrigatórios ausentes ou inválidos."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      },
      "get": {
        "tags": [
          "user"
        ],
        "description": "Endpoint para listar todos os usuários com paginação",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "per_page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "orderBy",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso"
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/user/{id}": {
      "put": {
        "tags": [
          "user"
        ],
        "description": "Endpoint para editar um usuário. Atualiza as informações de um usuário existente com base no ID fornecido.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "Informações necessárias para criar um novo usuário",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "user@example.com"
                },
                "nome": {
                  "type": "string",
                  "example": "João da Silva"
                },
                "senha": {
                  "type": "string",
                  "example": "senha123"
                },
                "permission": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário editado com sucesso"
          },
          "422": {
            "description": "Erro de validação. Campos obrigatórios ausentes ou inválidos."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      },
      "get": {
        "tags": [
          "user"
        ],
        "description": "Endpoint para buscar somente as informações de um usuário pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário retornado com sucesso"
          },
          "404": {
            "description": "Usuário não encontrado."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "description": "Endpoint para deletar um usuário pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário excluido com sucesso"
          },
          "404": {
            "description": "Usuário não encontrado."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/permission": {
      "get": {
        "tags": [
          "permission"
        ],
        "description": "Endpoint para listar todas as possivéis permissões (Listagem simples, sem paginação)",
        "responses": {
          "200": {
            "description": "Lista de permissões retornada com sucesso"
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    }
  }
}